refactor: move genre utils to utils/movie folder
